move readings type to its own header and source
