+++ nrg/src/readings_type.cpp
@@ -1,0 +1,32 @@
+#include <nrg/readings_type.hpp>
+
+#include <type_traits>
+
+namespace nrgprf::readings_type
+{
+    type operator|(type lhs, type rhs)
+    {
+        return static_cast<type>(
+            static_cast<std::underlying_type_t<type>>(lhs) |
+            static_cast<std::underlying_type_t<type>>(rhs)
+            );
+    }
+
+    type operator&(type lhs, type rhs)
+    {
+        return static_cast<type>(
+            static_cast<std::underlying_type_t<type>>(lhs) &
+            static_cast<std::underlying_type_t<type>>(rhs)
+            );
+    }
+
+    type operator^(type lhs, type rhs)
+    {
+        return static_cast<type>(
+            static_cast<std::underlying_type_t<type>>(lhs) ^
+            static_cast<std::underlying_type_t<type>>(rhs)
+            );
+    }
+
+    const type all = power | energy;
+}
